Rewrite the event migration and seeder
Why?
To eliminate event_groups table, there's no need for this, because for each event it has only one user who created it and associated with only one organization as well as event type and categories. So, with all this information, it should be in one table only, that is the event table.